Add files via upload
AddEditor, a GUI for modeling, is a molecular editing tool that adds several convenient features to the SurfaceEditor. For example, it allows you to extend molecular chains, modify with hydrogen, and perform structure optimizaion on edited molecules. These features enables the creation of new molecules and design of catalytic surfaces directly within the GUI.

Signed-off-by: SatoToshiya <107450221+tsyst@users.noreply.github.com>